- Adding TFv2 build and run resources (`run_TFv2_environment.sh`, `build_TFv2_environment.sh`, `Dockerfile-TFv2`) - Opening up port `6006` in docker network to support connections to tensorboard. - Renaming notebook assets for better sorting in notebook directory view. - Across almost all notebooks.   - Fix paths for container mode, changing `../../` to `../`

+++ mltf-entrypoint.sh
@@ -1,7 +1,9 @@
 #!/usr/bin/env bash
-
+set -e
 # Install Py2 kernel
-/usr/install/python27/bin/python -m ipykernel install --user
+if [[ -f /usr/install/python27/bin/python ]]; then
+    /usr/install/python27/bin/python -m ipykernel install --user
+fi
 
 # Start Jupyter and monitor it
 jupyter notebook --notebook-dir=/usr/src/mltf2 --no-browser --ip  0.0.0.0 --allow-root
